Mimic Rubycritic folder structure
- [x] Move command_factory file
- [x] Move commands folder
