--- lib/skunk/cli/commands/version.rb
@@ -1,22 +1,0 @@
-# frozen_string_literal: true
-
-require "rubycritic/commands/version"
-
-# nodoc #
-module Skunk
-  module Cli
-    module Command
-      # Shows skunk version
-      class Version < RubyCritic::Command::Version
-        def execute
-          print Skunk::VERSION
-          status_reporter
-        end
-
-        def sharing?
-          false
-        end
-      end
-    end
-  end
-end
