+++ lib/skunk/commands/output.rb
