+++ lib/skunk/commands/status_reporter.rb
