+++ lib/skunk/commands/version.rb
@@ -1,0 +1,20 @@
+# frozen_string_literal: true
+
+require "rubycritic/commands/version"
+
+# nodoc #
+module Skunk
+  module Command
+    # Shows skunk version
+    class Version < RubyCritic::Command::Version
+      def execute
+        print Skunk::VERSION
+        status_reporter
+      end
+
+      def sharing?
+        false
+      end
+    end
+  end
+end
