+++ test/lib/skunk/commands/version_test.rb
@@ -2,15 +2,15 @@
 
 require "test_helper"
 
-require "skunk/cli/commands/version"
+require "skunk/commands/version"
 require "skunk/cli/options"
 
-describe Skunk::Cli::Command::Version do
+describe Skunk::Command::Version do
   describe "#execute" do
     let(:msg) { Skunk::VERSION }
     let(:options) { ["--version"] }
     let(:opts) { Skunk::Cli::Options.new(options).parse }
-    let(:subject) { Skunk::Cli::Command::Version.new(opts.to_h) }
+    let(:subject) { Skunk::Command::Version.new(opts.to_h) }
 
     it "outputs the right version message" do
       assert_output(msg) do
